# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Preço Justo API",
    "version": "1.0.0",
    "description": "API do backend do Preço Justo (CRUD e operações principais)."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
...
    "/produtos": {
      "get": {
        "tags": ["Produtos"],
        "summary": "Listar produtos",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "in": "query",
            "name": "categoria_id",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "q", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista paginada" } }
      },
      "post": {
        "tags": ["Produtos"],
        "summary": "Criar produto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Produto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Criado" },
          "400": {
            "description": "Campos obrigatórios ausentes ou categoria inválida"
          }
        }
      }
    },
    "/produtos/{id}": {
      "get": {
        "tags": ["Produtos"],
        "summary": "Obter produto",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Não encontrado" }
        }
      },
      "put": {
        "tags": ["Produtos"],
        "summary": "Atualizar produto",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Produto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Atualizado" },
          "404": { "description": "Não encontrado" }
        }
      },
      "delete": {
        "tags": ["Produtos"],
        "summary": "Excluir produto",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Excluído" },
          "404": { "description": "Não encontrado" }
        }
      }
    },
    "/precos": {
      "get": {
        "tags": ["Preços"],
        "summary": "Listar preços",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "in": "query",
            "name": "produto_id",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "estabelecimento_id",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "usuario_id",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Lista paginada" } }
      },
      "post": {
        "tags": ["Preços"],
        "summary": "Criar preço",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Preco" }
            }
          }
        },
        "responses": {
          "201": { "description": "Criado" },
          "400": {
            "description": "Campos obrigatórios ausentes ou IDs inválidos"
          }
        }
      }
    },
...
    "/listas_compras": {
      "get": {
        "tags": ["Listas de Compras"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Listar listas",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "in": "query",
            "name": "usuario_id",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "ativa", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Lista paginada" } }
      },
      "post": {
        "tags": ["Listas de Compras"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Criar lista",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ListaCompras" }
            }
          }
        },
        "responses": {
          "201": { "description": "Criado" },
          "400": {
            "description": "Campos obrigatórios ausentes ou usuario_id inválido"
          }
        }
      }
    },
    "/listas_compras/{id}": {
      "get": {
        "tags": ["Listas de Compras"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Obter lista",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Não encontrado" }
        }
      },
      "put": {
        "tags": ["Listas de Compras"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Atualizar lista",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ListaCompras" }
            }
          }
        },
        "responses": {
          "200": { "description": "Atualizado" },
          "404": { "description": "Não encontrado" }
        }
      },
      "delete": {
        "tags": ["Listas de Compras"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Excluir lista",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Excluído" },
          "404": { "description": "Não encontrado" }
        }
      }
    },
...
    "/usuarios": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Listar usuários",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": { "200": { "description": "Lista paginada" } }
      },
      "post": {
        "tags": ["Usuários"],
        "summary": "Cadastrar usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Usuario" }
            }
          }
        },
        "responses": {
          "201": { "description": "Criado" },
          "400": { "description": "Campos obrigatórios ausentes" }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Obter usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Não encontrado" }
        }
      },
      "put": {
        "tags": ["Usuários"],
        "summary": "Atualizar usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Usuario" }
            }
          }
        },
        "responses": {
          "200": { "description": "Atualizado" },
          "404": { "description": "Não encontrado" }
        }
      },
      "delete": {
        "tags": ["Usuários"],
        "summary": "Excluir usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Excluído" },
          "404": { "description": "Não encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Produto": {
        "type": "object",
        "properties": {
          "produto_id": { "type": "string" },
          "nome": { "type": "string" },
          "nome_normalizado": { "type": "string" },
          "unidade": { "type": "string" },
          "categoria_id": { "type": "string" },
          "descricao": { "type": "string" },
          "imagem_url": { "type": "string" },
          "ean": { "type": "string" },
          "ativo": { "type": "boolean" }
        },
        "required": ["produto_id", "nome", "nome_normalizado", "unidade"]
      },
      "Preco": {
        "type": "object",
        "properties": {
          "preco_id": { "type": "string" },
          "produto_id": { "type": "string" },
          "estabelecimento_id": { "type": "string" },
          "usuario_id": { "type": "string" },
          "preco": { "type": "number" },
          "data_coleta": { "type": "string", "format": "date-time" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "validado": { "type": "boolean" }
        },
        "required": [
          "preco_id",
          "produto_id",
          "estabelecimento_id",
          "usuario_id",
          "preco"
        ]
      },
...
      "ListaCompras": {
        "type": "object",
        "properties": {
          "lista_compras_id": { "type": "string" },
          "usuario_id": { "type": "string" },
          "nome": { "type": "string" },
          "orcamento_total": { "type": "number" },
          "ativa": { "type": "boolean" }
        },
        "required": ["lista_compras_id", "usuario_id", "nome"]
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "usuario_id": { "type": "string" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "senha": { "type": "string" }
        },
        "required": ["usuario_id", "nome", "email", "senha"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}